rename resotolog -> resotoeventlog
