+++ resotoeventlog/MANIFEST.in
@@ -1,4 +1,4 @@
 include requirements-dev.txt
 include requirements-test.txt
 include requirements.txt
-include resotolog/ui/*
+include resotoeventlog/ui/*
