+++ resotoeventlog/logview/.gitignore
