+++ resotoeventlog/logview/public/favicon.ico
