+++ resotoeventlog/logview/public/index.html
