+++ resotoeventlog/logview/public/manifest.json
