+++ resotoeventlog/logview/src/App.js
