+++ resotoeventlog/logview/src/App.test.js
