+++ resotoeventlog/logview/src/react-app-env.d.ts
