+++ resotoeventlog/logview/tsconfig.json
