+++ resotoeventlog/requirements.txt
