+++ resotoeventlog/resotoeventlog/model.py
@@ -34,7 +34,7 @@ class LogConfig:
                 ctx.load_cert_chain(self.args.cert, self.args.cert_key, self.args.cert_key_pass)
                 return ctx
             else:
-                tls = TLSData(common_name="resotolog", resotocore_uri=self.core_uri.http_uri)
+                tls = TLSData(common_name="resotoeventlog", resotocore_uri=self.core_uri.http_uri)
                 ctx = create_default_context(Purpose.CLIENT_AUTH)
                 ctx.load_cert_chain(tls.cert_path, tls.key_path)
                 # TODO: renew ssl context
