+++ resotoeventlog/resotoeventlog/ui/manifest.json
