+++ resotoeventlog/resotoeventlog/ui/service-worker.js
