+++ resotoeventlog/resotoeventlog/web/directives.py
