+++ resotoeventlog/tox.ini
@@ -1,7 +1,7 @@
 [flake8]
 max-line-length=120
 exclude = .git,.tox,__pycache__,.idea,.pytest_cache
-application-import-names=resotolog tests
+application-import-names=resotoeventlog tests
 
 [pytest]
 testpaths= tests
@@ -16,7 +16,7 @@ deps =
    -rrequirements.txt
 
 [testenv:syntax]
-commands = flake8 --verbose resotolog
+commands = flake8 --verbose resotoeventlog
 
 [testenv:tests]
 commands = pytest
@@ -26,6 +26,6 @@ commands = black --line-length 120 --check --diff --target-version py39 .
 
 [testenv:mypy]
 commands=
-  - python -m mypy --python-version 3.9 resotolog tests
-  - python -m mypy --install-types --non-interactive --python-version 3.9 resotolog tests
-  python -m mypy --python-version 3.9 --strict resotolog tests
+  - python -m mypy --python-version 3.9 resotoeventlog tests
+  - python -m mypy --install-types --non-interactive --python-version 3.9 resotoeventlog tests
+  python -m mypy --python-version 3.9 --strict resotoeventlog tests
